basic refactor the logic of manager
